Move application settings to its own folder
