+++ Examples/Nodify.Workflow/App.xaml
@@ -9,7 +9,7 @@
             <ResourceDictionary.MergedDictionaries>
                 <ResourceDictionary Source="/Common/ControlStyles.xaml" />
                 <ResourceDictionary Source="/Common/CommonDataTemplates.xaml" />
-                <ResourceDictionary Source="/Shell/ApplicationSettingsDataTemplates.xaml" />
+                <ResourceDictionary Source="/Settings/ApplicationSettingsDataTemplates.xaml" />
             </ResourceDictionary.MergedDictionaries>
         </ResourceDictionary>
     </Application.Resources>
