+++ Examples/Nodify.Workflow/Settings/ApplicationSettingsDataTemplates.xaml
@@ -1,5 +1,5 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:local="clr-namespace:Nodify.Workflow.Shell"
+                    xmlns:local="clr-namespace:Nodify.Workflow.Settings"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
 
     <DataTemplate DataType="{x:Type local:KeybindingsSettingsCategoryViewModel}">
