+++ Examples/Nodify.Workflow/Settings/ApplicationSettingsWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="Nodify.Workflow.Shell.ApplicationSettingsWindow"
+﻿<Window x:Class="Nodify.Workflow.Settings.ApplicationSettingsWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Nodify.Workflow.Shell"
+        xmlns:local="clr-namespace:Nodify.Workflow.Settings"
         xmlns:ic="clr-namespace:FluentIcons.Wpf;assembly=FluentIcons.Wpf"
         d:DataContext="{d:DesignInstance local:ApplicationSettingsViewModel, IsDesignTimeCreatable=True}"
         WindowStyle="ToolWindow"
