+++ Examples/Nodify.Workflow/Settings/ApplicationSettingsWindow.xaml.cs
@@ -1,0 +1,14 @@
+﻿using System.Windows;
+
+namespace Nodify.Workflow.Settings;
+
+/// <summary>
+/// Interaction logic for ApplicationSettingsWindow.xaml
+/// </summary>
+public partial class ApplicationSettingsWindow : Window
+{
+    public ApplicationSettingsWindow()
+    {
+        InitializeComponent();
+    }
+}
