+++ Examples/Nodify.Workflow/Settings/KeybindingsCategory.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="Nodify.Workflow.Shell.KeybindingsCategory"
+﻿<UserControl x:Class="Nodify.Workflow.Settings.KeybindingsCategory"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
