--- Examples/Nodify.Workflow/Shell/ApplicationSettingsWindow.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System.Windows;
-
-namespace Nodify.Workflow.Shell
-{
-    /// <summary>
-    /// Interaction logic for ApplicationSettingsWindow.xaml
-    /// </summary>
-    public partial class ApplicationSettingsWindow : Window
-    {
-        public ApplicationSettingsWindow()
-        {
-            InitializeComponent();
-        }
-    }
-}
