+++ Examples/Nodify.Workflow/Shell/WorkflowsPanel.xaml.cs
@@ -1,15 +1,14 @@
 ﻿using System.Windows.Controls;
 
-namespace Nodify.Workflow.Shell
+namespace Nodify.Workflow.Shell;
+
+/// <summary>
+/// Interaction logic for WorkflowsPanel.xaml
+/// </summary>
+public partial class WorkflowsPanel : UserControl
 {
-    /// <summary>
-    /// Interaction logic for WorkflowsPanel.xaml
-    /// </summary>
-    public partial class WorkflowsPanel : UserControl
+    public WorkflowsPanel()
     {
-        public WorkflowsPanel()
-        {
-            InitializeComponent();
-        }
+        InitializeComponent();
     }
 }
